For some reason, the .pdfs in the library aren't correct.
I don't know why it's not updated correctly.
I think it has something to do with some of the strange default settings
in texlipse and the pdf build being a little behind the .tex file if
it's set as derived.

I'm pushing the updated .pdf for the SDP and the SVVS